fix display timestamp

+++ src/utils/renders.ts
@@ -4,7 +4,7 @@ export const displayTimestamp = (
 ) => {
 	const date = new Date(timestamp).toLocaleTimeString([], {
 		year: 'numeric',
-		month: 'numeric',
+		month: 'long',
 		day: 'numeric',
 		hour: '2-digit',
 		minute: '2-digit',
